requirements.txt
flask
pyyaml
pydantic
psycopg2-binary
